+++ api/src/main/scala/com/tersesystems/echopraxia/plusscala/api/ThrowableToName.scala
@@ -1,0 +1,9 @@
+package com.tersesystems.echopraxia.plusscala.api
+
+import com.tersesystems.echopraxia.spi.FieldConstants
+
+trait ThrowableToName {
+
+  // All exceptions should use "exception" field constant by default
+  implicit def throwableToName[T <: Throwable]: ToName[T] = ToName.create(FieldConstants.EXCEPTION)
+}
