+++ api/src/main/scala/com/tersesystems/echopraxia/plusscala/api/package.scala
@@ -1,3 +1,7 @@
 package com.tersesystems.echopraxia.plusscala
 
+import com.tersesystems.echopraxia.spi.FieldCreator
+import com.tersesystems.echopraxia.spi.EchopraxiaService
+import com.tersesystems.echopraxia.api.Field
+
 package object api extends LowPriorityImplicits
